requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
reportlab==4.1.0
xhtml2pdf==0.2.13
django-storages==1.14.2
boto3==1.34.34
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.57.0
django-widget-tweaks==1.5.0
django-tables2==2.6.0
django-bootstrap5==23.3
django-chartjs==2.3.0
django-notifications-hq==1.8.2
django-simple-history==3.4.0
django-import-export==3.3.4 